nft-reco-back/requirements.txt
# Base runtime
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
aiofiles==23.2.1
python-multipart==0.0.9

# Core ML stack
torch==2.2.2
torchvision==0.17.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0
openai-clip==1.0.1
transformers==4.37.2
clip-interrogator==0.6.0

# Feedback and hybrid search
scikit-learn==1.3.2
lightfm==1.17

# Visualization and evaluation
matplotlib==3.8.3
seaborn==0.13.1
pandas==2.2.1

# Utilities
numpy==1.26.4
pillow==10.3.0
typing-extensions==4.11.0
tqdm==4.66.1

# Testing
pytest==7.4.4

# Optional GPU build (uncomment if you have a compatible GPU)
# faiss-gpu==1.8.0
